completed examples
