+++ examples/with-nonce-manager/package.json
@@ -3,7 +3,7 @@
   "version": "0.1.3",
   "private": true,
   "scripts": {
-    "start-simple": "pnpm -w run build-all && tsx src/simple.ts",
+    "start-simple-sequential": "pnpm -w run build-all && tsx src/simpleSequential.ts",
     "start-managed-sequential": "pnpm -w run build-all && tsx src/managedSequential.ts",
     "start-managed-optimistic": "pnpm -w run build-all && tsx src/managedOptimistic.ts",
     "clean": "rimraf ./dist ./.cache",
